pyproject.toml
[project]
name = "carcare"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=5.2.7",
    "djangorestframework>=3.16.1",
    "djangorestframework-stubs>=3.16.4",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.1.1",
]
